+++ pinq/query/select.py
@@ -1,18 +1,18 @@
 
 from typing import Callable
 from typing_extensions import Self
-from pinq import enumerable_interface
+from pinq.enumerable_interface import IEnumerable
 from pinq.utils.lambda_utils import LambdaUtils
 from pinq.utils.decorators import extends
 
 
-@extends(enumerable_interface)
-def Select(self, func : Callable[..., object]) -> Self:
-    match LambdaUtils.get_number_of_arguments_taken(func):
+@extends(IEnumerable)
+def Select(self, selector : Callable[..., object]) -> IEnumerable:
+    match LambdaUtils.get_number_of_arguments_taken(selector):
         case 1:
-            self.generator = (func(value) for value in self.generator)
+            self.generator = (selector(value) for value in self.generator)
         case 2:
-            self.generator = (func(value, index) for value, index in enumerate(self.generator))
+            self.generator = (selector(value, index) for value, index in enumerate(self.generator))
         case _:
-            raise ValueError("Callable must take one or two arguments")
+            raise ValueError("Given function must take one or two arguments")
     return self